+++ dhg/structure/__init__.py
@@ -1,7 +1,7 @@
 from .base import load_structure, BaseGraph, BaseHypergraph
 from .graphs import Graph, DiGraph, BiGraph
 from .hypergraphs import Hypergraph
-
+from .hypergraphs import MultilayerHypergraph
 
 __all__ = [
     "load_structure",
@@ -11,4 +11,5 @@ __all__ = [
     "DiGraph",
     "BiGraph",
     "Hypergraph",
+    "MultilayerHypergraph"
 ]
