+++ dhg/structure/hypergraphs/__init__.py
@@ -1,1 +1,2 @@
 from .hypergraph import Hypergraph
+from .multilayer_hypergraph import MultilayerHypergraph
